moved ApiErr into api_common
