+++ src/api_480.rs
@@ -18,10 +18,10 @@ pub use crate::internal::ByteVector;
 use crate::nonemptyvec::NonEmptyVec;
 use clear_on_drop::clear::Clear;
 use gridiron::fp_480::Fp480;
-use quick_error::quick_error;
 use rand;
 use std;
 use std::fmt;
+use crate::api_common::{ApiErr, Result};
 
 /// Recrypt public API
 #[derive(Debug)]
@@ -55,52 +55,6 @@ impl<CR: rand::CryptoRng + rand::RngCore> Api<Sha256, Ed25519, RandomBytes<CR>> 
     }
 }
 
-quick_error! {
-    /// Errors generated by the API
-    #[derive(Debug, PartialEq)]
-    pub enum ApiErr {
-        DecryptFailed(err: internal::InternalError){
-            display("The decryption failed. Ensure you're using the correct PrivateKey.")
-            cause(err)
-        }
-        InvalidEncryptedMessageSignature(err: internal::InternalError){
-            display("The signature of the encrypted value could not be verified.")
-            cause(err)
-        }
-        InvalidPublicKey(err: internal::homogeneouspoint::PointErr){
-            display("The public key was not valid. Ensure it was generated from a valid PrivateKey.")
-            from()
-            cause(err)
-        }
-        InvalidTransformKey(err: internal::InternalError){
-            cause(err)
-            display("The transform key signature was incorrect.")
-        }
-        InputWrongSize(typ: &'static str, req_size: usize){
-            display("The input value was the wrong size. Expected {} bytes for type {}.", req_size, typ)
-        }
-        DecodeFailure(err: internal::bytedecoder::DecodeErr){
-            display("The bytes could not be decoded into the appropriate data type.")
-            cause(err)
-            from()
-        }
-    }
-}
-
-type Result<T> = std::result::Result<T, ApiErr>;
-
-impl From<internal::InternalError> for ApiErr {
-    fn from(err: internal::InternalError) -> Self {
-        match err {
-            internal::InternalError::AuthHashMatchFailed => ApiErr::DecryptFailed(err),
-            internal::InternalError::InvalidEncryptedMessageSignature => {
-                ApiErr::InvalidEncryptedMessageSignature(err)
-            }
-            internal::InternalError::PointInvalid(p_err) => ApiErr::InvalidPublicKey(p_err),
-            internal::InternalError::CorruptReencryptionKey => ApiErr::InvalidTransformKey(err),
-        }
-    }
-}
 
 /// Hashed but not encrypted Plaintext used for envelope encryption
 new_bytes_type!(DerivedSymmetricKey, 32);
@@ -1118,13 +1072,6 @@ pub(crate) mod test {
     use crate::internal::fp::fp480_unsafe_from;
     use hex;
     use rand_chacha;
-
-    //Writing a BS from conversion to make tests which contain `hex` constants easier to write.
-    impl From<hex::FromHexError> for ApiErr {
-        fn from(err: hex::FromHexError) -> Self {
-            panic!("Error '{:?}' was found when parsing hex in tests.", err);
-        }
-    }
 
     pub struct DummyEd25519;
     impl Ed25519Signing for DummyEd25519 {
